// SpecialEducationPlanning/SpecialEducationPlanning.Web/src/app/shared/models/plan.ts
import { Project } from './project';
import { Version } from './version';
import { EndUser } from './end-user';
import { EducationToolType, PlanStateEnum } from './app-enums';
import { UserEducationer } from './Educationer.model';
import { Builder } from './builder';
import { HousingType } from './housing-type.model';
import { HousingSpecificationTemplates } from './housing-specification-templates.model';
import { ProjectTemplates } from './project-templates.model';
export interface Plan {
  id: number;
  cadFilePlanId?: string | null;
  title?: string | null;
  lastOpen?: Date | string;
  updatedDate: Date | string;
  updateUser?: string | null;
  createdDate: Date | string;
  creationUser?: string | null;
  catalogId: number;
  EducationerId?: number | null;
  Educationer: UserEducationer;
  projectId: number;
  project?: Project;
  versions: Version[];
  keyName?: string | null;
  survey: boolean;
  planCode?: string | null;
  endUser?: EndUser;
  endUserId?: number | null;
  endUserFullName?: string;
  planState: PlanStateEnum;
  builderId?: number | null;
  builder?: Builder;
  masterVersionId?: number | null;
  builderTradingName?: string | null;
  planType?: string | null;
  notes: string;
  masterVersion?: Version;
  planName?: string | null;
  offlineSyncDate?: Date | string | null;
  isStarred: boolean;
  isTemplate?: boolean;
  housingTypeId?: number | null;
  housingTypeModel?: HousingType;
  housingSpecificationTemplatesId?: number | null;
  housingSpecificationTemplatesModel?: HousingSpecificationTemplates;
  projectTemplatesId?: number | null;
  projectTemplatesModel?: ProjectTemplates;
  housingSpecificationId?: number | null;
  planReference?: string;
  EducationOrigin?: EducationToolType;
}